googleAPI/influencerChat/requirements.txt
streamlit
google-generativeai
chromadb
python-dotenv
tqdm
ipython